Moves common code from WebsocketStreamStubs into an abstract base class.
